Crate allegro

Reexports

pub use allegro_util::*;

Macros

allegro_main!

Structs

Bitmap

BitmapDrawingFlags

BitmapFlags

Color

Config
Allegro configuration.

ConfigEntry
Configuration entry. Used for iterating through all the entries of a configuration.

ConfigSection
Configuration section. Used for iterating through all the sections of a configuration.

Core

Display

DisplayFlags

EventQueue

EventSource

Joystick

KeyModifier

MemoryBitmap
A memory bitmap, unlike a regular bitmap, can be transferred between threads.

Shader
A shader program comprising of a pixel (fragment) and a vertex shader. Wraps ALLEGRO_SHADER.

StickFlags

SubBitmap

Timer

Transform

Enums

DisplayOption

DisplayOptionImportance

DisplayOrientation

Event

KeyCode

PixelFormat

ShaderPlatform
Shader platform.

ShaderType
Shader type.

Constants

ACCENT1

ACCENT2

ACCENT3

ACCENT4

ALPHA_TEST

ALT

ALTGR

ANALOGUE

CAPSLOCK

COMMAND

CTRL

DIGITAL

DIRECT3D

FLIP_HORIZONTAL

FLIP_NONE

FLIP_VERTICAL

FORCE_LOCKING

FRAMELESS

FULLSCREEN

FULLSCREEN_WINDOW

GENERATE_EXPOSE_EVENTS

INALTSEQ

KEEP_BITMAP_FORMAT

LWIN

MAG_LINEAR

MEMORY_BITMAP

MENU

MINIMIZED

MIN_LINEAR

MIPMAP

NO_PREMULTIPLIED_ALPHA

NO_PRESERVE_TEXTURE

NUMLOCK

OPENGL

OPENGL_3_0

OPENGL_FORWARD_COMPATIBLE

RESIZABLE

RWIN

SCROLLLOCK

SHIFT

VIDEO_BITMAP

WINDOWED

Traits

BitmapLike
A trait implemented by types that behave like bitmaps.

SharedBitmap
Expresses something that shares access with a real bitmap.

Functions

run