Struct allegro::core::Core

pub struct Core { /* fields omitted */ }

Methods

impl Core

fn init() -> Result<Core, String>

This must be called on the main thread.

fn get_system_config() -> Config

Returns the system config. TODO: This isn't quite thread safe...

fn spawn<F: FnOnce(Core) + Send + 'static>(&self, thread_proc: F)

fn get_core_mutex(&self) -> Arc<Mutex<()>>

fn get_num_video_adapters(&self) -> i32

fn get_monitor_info(&self, adapter: i32) -> Result<(i32, i32, i32, i32), ()>

fn rest(&self, seconds: f64)

fn get_time(&self) -> f64

fn install_keyboard(&self) -> Result<(), ()>

fn is_keyboard_installed(&self) -> bool

fn get_keyboard_event_source(&self) -> Option<EventSource>

fn set_keyboard_leds(&self, leds: KeyModifier) -> Result<(), ()>

fn keycode_to_name(&self, k: KeyCode) -> String

fn install_mouse(&self) -> Result<(), ()>

fn is_mouse_installed(&self) -> bool

fn get_mouse_event_source(&self) -> Option<EventSource>

fn install_joystick(&self) -> Result<(), ()>

fn is_joystick_installed(&self) -> bool

fn get_joystick_event_source(&self) -> Option<EventSource>

fn reconfigure_joysticks(&self) -> Result<(), ()>

fn get_num_joysticks(&self) -> i32

fn get_mouse_num_buttons(&self) -> u32

fn get_mouse_num_axes(&self) -> u32

fn set_mouse_xy(&self, display: &Display, x: i32, y: i32) -> Result<(), ()>

fn set_mouse_z(&self, z: i32) -> Result<(), ()>

fn set_mouse_w(&self, w: i32) -> Result<(), ()>

fn set_mouse_axis(&self, axis: i32, value: i32) -> Result<(), ()>

fn grab_mouse(&self, display: &Display) -> Result<(), ()>

fn ungrab_mouse(&self) -> Result<(), ()>

fn set_new_bitmap_flags(&self, flags: BitmapFlags)

fn get_new_bitmap_flags(&self) -> BitmapFlags

fn set_new_bitmap_format(&self, format: PixelFormat)

fn get_new_bitmap_format(&self) -> PixelFormat

fn set_target_bitmap<T: BitmapLike>(&self, bmp: &T)

fn clear_to_color(&self, color: Color)

fn draw_pixel(&self, x: f32, y: f32, color: Color)

fn put_pixel(&self, x: i32, y: i32, color: Color)

fn put_blended_pixel(&self, x: i32, y: i32, color: Color)

fn draw_bitmap<T: BitmapLike>(
&self,
bitmap: &T,
dx: f32,
dy: f32,
flags: BitmapDrawingFlags
)

fn draw_bitmap_region<T: BitmapLike>(
&self,
bitmap: &T,
sx: f32,
sy: f32,
sw: f32,
sh: f32,
dx: f32,
dy: f32,
flags: BitmapDrawingFlags
)

fn draw_scaled_bitmap<T: BitmapLike>(
&self,
bitmap: &T,
sx: f32,
sy: f32,
sw: f32,
sh: f32,
dx: f32,
dy: f32,
dw: f32,
dh: f32,
flags: BitmapDrawingFlags
)

fn draw_rotated_bitmap<T: BitmapLike>(
&self,
bitmap: &T,
cx: f32,
cy: f32,
dx: f32,
dy: f32,
angle: f32,
flags: BitmapDrawingFlags
)

fn draw_scaled_rotated_bitmap<T: BitmapLike>(
&self,
bitmap: &T,
cx: f32,
cy: f32,
dx: f32,
dy: f32,
xscale: f32,
yscale: f32,
angle: f32,
flags: BitmapDrawingFlags
)

fn draw_tinted_bitmap<T: BitmapLike>(
&self,
bitmap: &T,
tint: Color,
dx: f32,
dy: f32,
flags: BitmapDrawingFlags
)

fn draw_tinted_bitmap_region<T: BitmapLike>(
&self,
bitmap: &T,
tint: Color,
sx: f32,
sy: f32,
sw: f32,
sh: f32,
dx: f32,
dy: f32,
flags: BitmapDrawingFlags
)

fn draw_tinted_scaled_bitmap<T: BitmapLike>(
&self,
bitmap: &T,
tint: Color,
sx: f32,
sy: f32,
sw: f32,
sh: f32,
dx: f32,
dy: f32,
dw: f32,
dh: f32,
flags: BitmapDrawingFlags
)

fn draw_tinted_rotated_bitmap<T: BitmapLike>(
&self,
bitmap: &T,
tint: Color,
cx: f32,
cy: f32,
dx: f32,
dy: f32,
angle: f32,
flags: BitmapDrawingFlags
)

fn draw_tinted_scaled_rotated_bitmap<T: BitmapLike>(
&self,
bitmap: &T,
tint: Color,
cx: f32,
cy: f32,
dx: f32,
dy: f32,
xscale: f32,
yscale: f32,
angle: f32,
flags: BitmapDrawingFlags
)

fn draw_tinted_scaled_rotated_bitmap_region<T: BitmapLike>(
&self,
bitmap: &T,
sx: f32,
sy: f32,
sw: f32,
sh: f32,
tint: Color,
cx: f32,
cy: f32,
dx: f32,
dy: f32,
xscale: f32,
yscale: f32,
angle: f32,
flags: BitmapDrawingFlags
)

fn set_clipping_rectangle(&self, x: i32, y: i32, width: i32, height: i32)

fn reset_clipping_rectangle(&self)

fn get_clipping_rectangle(&self) -> (i32, i32, i32, i32)

fn set_new_display_flags(&self, flags: DisplayFlags)

fn get_new_display_flags(&self) -> DisplayFlags

fn set_new_display_refresh_rate(&self, rate: i32)

fn get_new_display_refresh_rate(&self) -> i32

fn set_new_display_adapter(&self, adapter: i32)

fn get_new_display_adapter(&self) -> i32

fn set_new_window_position(&self, x: i32, y: i32)

fn get_new_window_position(&self) -> (i32, i32)

fn reset_new_display_options(&self)

fn set_new_display_option(
&self,
option: DisplayOption,
value: i32,
importance: DisplayOptionImportance
)

fn get_new_display_option(
&self,
option: DisplayOption
) -> (i32, DisplayOptionImportance)

fn get_current_transform(&self) -> Transform

fn use_transform(&self, trans: &Transform)

fn use_shader(&self, shader: Option<&Shader>) -> Result<(), ()>

Set the shader as current for the current bitmap. Pass None to stop using this shader. Returns an error if the shader isn't compatible with the bitmap.

fn get_default_shader_source(
&self,
platform: ShaderPlatform,
shader_type: ShaderType
) -> Option<String>

Returns the source of the shader that Allegro uses by default.

fn flip_display(&self)

fn update_display_region(&self, x: i32, y: i32, width: i32, height: i32)

fn wait_for_vsync(&self) -> Result<(), ()>

fn hold_bitmap_drawing(&self, hold: bool)

fn is_bitmap_drawing_held(&self) -> bool

fn set_shader_sampler<T: BitmapLike>(
&mut self,
name: &str,
bmp: &T,
unit: i32
) -> Result<(), ()>

Set a sampler for a particular uniform and unit for the current shader. Different uniforms should be set to different units. Pass None to bmp to clear the sampler.

fn set_shader_uniform<T: ShaderUniform + ?Sized>(
&self,
name: &str,
val: &T
) -> Result<(), ()>

Sets a shader uniform to a value.

fn set_blender(&self, op: BlendOperation, source: BlendMode, dest: BlendMode)

Set blender options.