Struct allegro::display::Display

pub struct Display { /* fields omitted */ }

Methods

impl Display

pub fn new(_: &Core, w: i32, h: i32) -> Result<Display, ()>

pub fn get_width(&self) -> i32

pub fn get_height(&self) -> i32

pub fn get_format(&self) -> PixelFormat

pub fn get_refresh_rate(&self) -> i32

pub fn get_flags(&self) -> DisplayFlags

pub fn set_flag(&self, flag: DisplayFlags, onoff: bool) -> bool

pub fn get_backbuffer(&self) -> &Bitmap

pub fn acknowledge_resize(&self) -> Result<(), ()>

pub fn resize(&self, w: i32, h: i32) -> Result<(), ()>

pub fn set_icon<T: BitmapLike>(&self, icon: &T)

pub fn set_icons<'l, U: Iterator<Item = &'l (BitmapLike + 'l)>>(&self, icons: U)

pub fn set_window_position(&self, x: i32, y: i32)

pub fn get_window_position(&self) -> (i32, i32)

pub fn set_window_title(&self, title: &str)

pub fn get_option(&self, option: DisplayOption) -> i32

pub fn convert_bitmap<T: BitmapLike>(&self, bmp: &T) -> Result<Bitmap, ()>

pub fn get_event_source(&self) -> EventSource

pub fn get_allegro_display(&self) -> *mut ALLEGRO_DISPLAY

pub fn create_shader(
&mut self,
platform: ShaderPlatform
) -> Result<Weak<Shader>, ()>

Create a new shader associated with this display.

Note that display destruction will panic if a strong reference is held to a shader at that time.