Struct allegro::Shader

pub struct Shader {
    // some fields omitted
}

A shader program comprising of a pixel (fragment) and a vertex shader. Wraps ALLEGRO_SHADER.

Methods

impl Shader

fn new(display: &mut Display, platform: ShaderPlatform) -> Result<Shader, ()>

Create a new shader for a particular platform.

A shader gets invalidated when the display is destroyed.

fn get_allegro_shader(&self) -> *mut ALLEGRO_SHADER

Returns the wrapped Allegro shader pointer.

fn attach_shader_source(&mut self, shader_type: ShaderType, source: Option<&str>) -> Result<(), ()>

Attach a source to the shader. Passing None clears the source.

fn attach_shader_source_file(&mut self, shader_type: ShaderType, filename: &str) -> Result<(), ()>

Attach a source to the shader that is loaded from a file.

fn build(&mut self) -> Result<(), ()>

Build the shader. Call this after attaching the sources.

fn get_log(&self) -> String

Get the log from the shader. Call this function if any of the attach/build functions fail to determine what went wrong.

fn get_platform(&self) -> Result<ShaderPlatform, ()>

Returns the platform of this shader.

fn set_sampler<T: BitmapLike>(&mut self, name: &str, bmp: &T, unit: i32) -> Result<(), ()>

Sets a sampler for a particular uniform and unit. Different uniforms should be set to different units. Pass None to bmp to clear the sampler.

fn is_valid(&self) -> bool