Struct allegro::config::Config

pub struct Config { /* fields omitted */ }

Allegro configuration.

Wraps ALLEGRO_CONFIG.

Methods

impl Config

pub fn new() -> Config

Creates an empty configuration.

pub unsafe fn wrap(config: *mut ALLEGRO_CONFIG, own: bool) -> Config

pub fn load(path: &str) -> Result<Config, ()>

Loads a config from a path.

pub fn merge(cfg1: &Config, cfg2: &Config) -> Config

Merge two configs into 1.

pub fn get_allegro_config(&self) -> *mut ALLEGRO_CONFIG

Returns the wrapped ALLEGRO_CONFIG.

pub fn save(&self, path: &str) -> Result<(), ()>

Loads a config from a path.

pub fn add_section(&mut self, name: &str)

Adds an empty section with the specified name.

pub fn remove_section(&mut self, name: &str) -> bool

Removes a section with the specified name.

pub fn set_value(&mut self, section: &str, key: &str, value: &str)

Sets the value of a key in a section.

pub fn remove_key(&mut self, section: &str, key: &str) -> bool

Removes a key with the specified name.

pub fn add_comment(&mut self, section: &str, comment: &str)

Adds a comment to a section.

pub fn get_value(&self, section: &str, key: &str) -> Option<String>

Gets a value from a section.

pub fn merge_from(&mut self, source: &Config)

Merge in sections from a different config.

Important traits for ConfigSection<'l>
impl<'l> Iterator for ConfigSection<'l> type Item = String;

pub fn sections<'l>(&'l self) -> ConfigSection<'l>

Returns an iterator over all the sections in the config. The first returned section will typically be the root section, even if it's empty.

Important traits for ConfigEntry<'l>
impl<'l> Iterator for ConfigEntry<'l> type Item = String;

pub fn keys<'l>(&'l self, section: &str) -> ConfigEntry<'l>

Returns an iterator over all the entries in a particular section.