Struct allegro::display::Display

pub struct Display { /* fields omitted */ }

Methods

impl Display

fn new(_: &Core, w: i32, h: i32) -> Result<Display, ()>

fn get_width(&self) -> i32

fn get_height(&self) -> i32

fn get_format(&self) -> PixelFormat

fn get_refresh_rate(&self) -> i32

fn get_flags(&self) -> DisplayFlags

fn set_flag(&self, flag: DisplayFlags, onoff: bool) -> bool

fn get_backbuffer(&self) -> &Bitmap

fn acknowledge_resize(&self) -> Result<(), ()>

fn resize(&self, w: i32, h: i32) -> Result<(), ()>

fn set_icon<T: BitmapLike>(&self, icon: &T)

fn set_icons<'l, U: Iterator<Item = &'l (BitmapLike + 'l)>>(&self, icons: U)

fn set_window_position(&self, x: i32, y: i32)

fn get_window_position(&self) -> (i32, i32)

fn set_window_title(&self, title: &str)

fn get_option(&self, option: DisplayOption) -> i32

fn convert_bitmap<T: BitmapLike>(&self, bmp: &T) -> Result<Bitmap, ()>

fn get_event_source(&self) -> EventSource

fn get_allegro_display(&self) -> *mut ALLEGRO_DISPLAY

fn create_shader(
&mut self,
platform: ShaderPlatform
) -> Result<Weak<Shader>, ()>

Create a new shader associated with this display.

Note that display destruction will panic if a strong reference is held to a shader at that time.